Transactions

+++ src/main/java/com/epam/controller/PaymentController.java
@@ -3,5 +3,5 @@ package com.epam.controller;
 import com.epam.exceptions.InsufficientFundsException;
 
 public interface PaymentController {
-    void deposit(Long userIdSender, Long userIdReceiver,  Long amount) throws InsufficientFundsException;
+    void deposit(Integer userIdSender, Integer userIdReceiver,  Long amount) throws InsufficientFundsException;
 }
