+++ src/main/java/com/epam/service/DepositService.java
@@ -4,5 +4,5 @@ import com.epam.exceptions.InsufficientFundsException;
 import com.epam.exceptions.UserNotFoundException;
 
 public interface DepositService {
-    boolean deposit(Long userIdSender, Long userIdReceiver,  Long amount) throws InsufficientFundsException, UserNotFoundException;
+    boolean deposit(Integer userIdSender, Integer userIdReceiver,  Long amount) throws InsufficientFundsException, UserNotFoundException;
 }
